+++ libpok/core/threadsuspend.c
@@ -29,7 +29,7 @@
 # if defined (POK_NEEDS_ARINC653)
 pok_ret_t pok_thread_suspend_other (uint32_t thread_id)
 {
-	return pok_syscall2  (POK_SYSCALL_THREAD_SUSPEND_OTHER, &thread_id, 0);
+	return pok_syscall2  (POK_SYSCALL_THREAD_SUSPEND_OTHER, (uint32_t)&thread_id, 0);
 }
 
 #endif
